Minor changes to get both Manticore CLI and script-based example via basic_statemerging.py to work + moving Merger Plugin to plugin.py
These changes now allow state merging to correctly merge two states in basic_statemerging.c
